// src/models/MyCollection.d.ts
export interface MyCollection{
    artist_name_kor:string;
    artist_name_eng:string;
    birth:string;
    death:string;
    user_art_id : number;
    user_id : number;
    artist_id : number;
    title_kor : string;
    title_eng : string;
    unit_cd : string;
    size_length : string;
    size_height : string;
    canvas: string;
    edition:string;
    image_name:null|string;
    create_time : string;
    price : number;
    buy_date:string;
    img_list:string[];
}